moved stuff into plugins
