+++ index.js
@@ -19,30 +19,37 @@ function checkHapiPluginError(error) {
 }
 
 // Register plugins
-const plugins = [{register: Chairo}, {register: Jwt}];
+const plugins = [
+  {
+    register: Chairo
+  },
+  {
+    register: Jwt
+  },
+  {
+    register: require('./lib/local_jwt_auth.js'),
+    options : Options
+  },
+  {
+    register: require('./routes/auth'),
+    options: Options
+  },
+  {
+    register: require('./routes/exercises')
+  },
+  {
+    register: require('./routes/calendar')
+  }
+];
 
 server.register(plugins, (err) => {
   checkHapiPluginError(err);
   const seneca = server.seneca;
 
   seneca.ready((err) => {
-
-    server.auth.strategy('jwt', 'jwt', {
-      key: Options.jwtKey,
-      validateFunc: authenticate
-    });
-
-    server.register(require('./routes/exercises'), (err) => {
-      checkHapiPluginError(err);
-    });
-
-    server.register(require('./routes/calendar'), (err) => {
-      checkHapiPluginError(err);
-    });
-
-    server.register(require('./routes/auth'), (err) => {
-      checkHapiPluginError(err);
-    });
+    if (err) {
+      throw err;
+    }
 
     Options.clients.forEach(function(opts) {
       console.log('Registering Client', JSON.stringify(opts))
@@ -54,16 +61,3 @@ server.register(plugins, (err) => {
     });
   }); 
 });
-
-function authenticate(decoded, request, cb) {
-  // console.log(request)
-  console.log('Decoded', decoded)
-  const token = decoded;
-  request.seneca.act({role: 'user', cmd:'auth', token: token}, (err, resp) => {
-    if (err) return cb(null, err);
-    if (resp.ok === false) {
-      return cb(null, false)
-    }
-    return cb(null, true)
-  });
-}
